Fixing the float saletax issue

+++ src/schema/mutations/index.ts
@@ -13,7 +13,7 @@ export const mutations = `
       """
       Update user
       """
-      updateUser(storeName: String, walletAddress: String, subDomain: String, currency: String, image: Upload, saleTax: Int): Store
+      updateUser(storeName: String, walletAddress: String, subDomain: String, currency: String, image: Upload, saleTax: Float): Store
 
       """
       Add Product
